backend/requirements.txt
Pillow
numpy
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
python-multipart
opencv-python-headless
openai
tensorflow-cpu
huggingface_hub
h5py
